Add support for extra configuration options
This adds an environment variable SQUCUMBER_OPTIONS_EXTRA that, if set,
allows to add extra configuration parameters to the underlying Cucumber
run command. It is appended to the existing configuration. This is
useful for adding e.g. `--strict` to let tests fail if steps are
undefined or for specifying additional output formats and directories.