Keep raw value if it has not been parsed
Whenever a value with parentheses would be encountered, the match groups
would be removed from the resulting string. This does not make sense as
the only case we're interested in a parsed value is when we're dealing
with date placeholders. This commit fixes the issue by returning the
original value if no date conversion has taken place.

Fixes #13