Fuzzy match today and yesterday
In case of e.g. using now() or current_timestamp() in the query, it's
hard to determine the exact hour, minute and seconds of execution for
the comparison. In these cases, using a fuzzy matcher such as "sometime
today" should be enough to ensure that the query works as designed,
without adding too much overhead for e.g. mocking out core functions of
the database.