Raise error for unknown column names
When specifying a non-existing column name in the expected result set,
this would be silently ignored. This would be confusing in case of typos
because they could lead to tests succeeding even in the face of errors.

This commit adds some additional checks in order to provide a more
meaningful error message when the returned result is empty, and to flag
typos immediately.

Fixes #18