Prevent system tables from being truncated
This would manifest as a permission denied error whenever there are more
than two scenarios in a feature and the Postgres has only a single
database object.

From local inspection, only pg_catalog and information_schema show up in
this query, although it might be wise in the future to exclude all pg_
prefixed schemas.