Ensure compatibility with data placeholders
In case of date placeholders being used in addition to NULL mappers,
we'd run into the issue of Ruby Nil values being tried to match with
regular expressions. This would cause some ugly exceptions in the code
and unnecessarily cause tests to fail.

This commit moves null mappings to the end of the mock value
conversions, so we know that date placeholders are fine.